+++ lib/ffmpeg/filter_graph/graph.rb
@@ -3,15 +3,15 @@ module FFmpeg::FilterGraph
     attr_accessor :chains, :outputs
 
     def initialize(chains: [], outputs: [])
-      self.chains = chains
-      self.outputs = outputs.map { |pad| Pad.out(pad) }
+      self.chains = Array(chains)
+      self.outputs = Array(outputs).map { |pad| Pad.out(pad) }
     end
 
     def to_s
       [
         chains.map(&:to_s).join('; '),
         outputs.map(&:to_s).join(''),
-      ].join('')
+      ].join(' ').strip
     end
   end
 end