+++ lib/ffmpeg/filter_graph/pad.rb
@@ -4,25 +4,22 @@ module FFmpeg::FilterGraph
 
     def self.in(pad)
       case pad
+      when Pads::Outpad then fail ArgumentError, 'cannot use Outpad as input'
       when Pad then pad
-      else
-        if /\A\d/ =~ pad.to_s
-          Pads::Input.new(pad.to_s)
-        else
-          Pads::Inpad.new(pad.to_s)
-        end
+      else Pads::Inpad.new(pad.to_s)
       end
     end
 
     def self.out(pad)
       case pad
+      when Pads::Inpad then fail ArgumentError, 'cannot use Inpad as output'
       when Pad then pad
       else Pads::Outpad.new(pad.to_s)
       end
     end
 
     def initialize(name)
-      fail ArgumentError, 'name cannot be nil' if name.nil?
+      fail ArgumentError, 'name cannot be empty' if name.nil? || name.empty?
 
       @name = name
     end