+++ lib/ffmpeg/filter_graph/pads/input.rb
@@ -1,9 +1,0 @@
-module FFmpeg::FilterGraph
-  module Pads
-    class Input < Pad
-      def to_s
-        name
-      end
-    end
-  end
-end