+++ test/ffmpeg/filter_graph/pads/inpad_test.rb
@@ -1,0 +1,12 @@
+require 'test_helper'
+
+module FFmpeg::FilterGraph
+  module Pads
+    class InpadTest < UnitTest
+      def test_to_s
+        assert_equal '[input]', Inpad.new('input').to_s
+      end
+    end
+  end
+end
+