+++ CHANGELOG.md
@@ -10,3 +10,5 @@
    underscore: `:_0a` -> `0a`
  - FilterGraph: Option names that include a dash can be specified in camelCase:
    `:softKnee` -> `soft-knee`
+ - Moved `Helper` mixin to `FFmpeg::FilterGraph::Utils`
+ - Inpads/Outpads are encapsulated in classes, instead of naked strings